requirements.txt
amqp==2.5.0
asn1crypto==0.24.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.3
django-allauth==0.39.1
django-cors-headers==3.0.2
django-heroku==0.3.1
django-rest-api==0.1.5
django-rest-auth==0.9.5
djangorestframework==3.10.1
gunicorn==19.9.0
idna==2.8
kombu==4.6.3
ndg-httpsclient==0.5.1
oauthlib==3.0.2
psycopg2==2.7.6.1
pusher==2.1.3
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
pyOpenSSL==19.0.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
whitenoise==4.1.3
